refactor: update RimeWithWeasel.cpp, style holded by session (#1055)
* refactor: update RimeWithWeasel.cpp, style holded by session

fix: style incorrect in some case when dark mode switches.

chore: UI function de-warning.

* feat: global ascii with option `global_ascii: {bool}`

* feat: `show_notifications_time: {int}` to determine how long a time to show notifications, unit: ms

* refactor show_notifications:

keyword: keyword `show_notifications_when` changed to `show_notifications`
type: show_notifications by bool/list of options, or `schema`

feat: able to show notifications of switches customized by user.

+++ include/WeaselUI.h
@@ -61,7 +61,7 @@ namespace weasel
 		// 界面显隐
 		void Show();
 		void Hide();
-		void ShowWithTimeout(DWORD millisec);
+		void ShowWithTimeout(size_t millisec);
 		bool IsCountingDown() const;
 		bool IsShown() const;
